refactor(transforms): move Standalize class to "common" module
