+++ src/sample/transforms/image.py
@@ -15,6 +15,8 @@ from torchvision.transforms.v2 import (
 
 from pamiq_vrchat.sensors.image import ImageFrame
 
+from .common import Standardize
+
 
 class ResizeAndCenterCrop(nn.Module):
     """Resize and center crop transform for images.
@@ -68,38 +70,6 @@ class ResizeAndCenterCrop(nn.Module):
         input = F.resize(input, list(scale_size))
         input = F.center_crop(input, self.size)
         return input
-
-
-class Standardize(nn.Module):
-    """Standardize input by removing mean and dividing by standard deviation.
-
-    This module performs standardization (zero mean, unit variance) on
-    the input tensor.
-    """
-
-    def __init__(self, eps: float = 1e-8) -> None:
-        """Initialize the Standardize transform.
-
-        Args:
-            eps: Small value to add to standard deviation to avoid division by zero.
-        """
-        super().__init__()
-        self.eps = eps
-
-    @override
-    def forward(self, input: Tensor) -> Tensor:
-        """Apply standardization to the input tensor.
-
-        Args:
-            input: Input tensor to standardize.
-
-        Returns:
-            Standardized tensor with zero mean and unit variance.
-        """
-        if input.numel() <= 1:  # Can not compute std
-            return torch.zeros_like(input)
-
-        return (input - input.mean()) / (input.std() + self.eps)
 
 
 def create_vrchat_transform(
